+++ src/rules/media-feature-value-dollar-variable/index.js
@@ -1,11 +1,11 @@
 import { utils } from "stylelint"
 import { namespace } from "../../utils"
 
-export const ruleName = namespace("media-query-value-dollar-variable")
+export const ruleName = namespace("media-feature-value-dollar-variable")
 
 export const messages = utils.ruleMessages(ruleName, {
-  rejected: "Unexpected dollar-variable as a value in a media query",
-  expected: "Expected a dollar-variable (e.g. $var) to be used as a value in a media query",
+  rejected: "Unexpected dollar-variable as a media feature value",
+  expected: "Expected a dollar-variable (e.g. $var) to be used as a media feature value",
 })
 
 export default function (expectation, options) {